Saving state
Current state:
1) dynamic collision detection working: if a new obstacle is published
in the environment, then the current solution paths will be rechecked
for validity and removed if they are in collision with the new obstacle.
2) dynamic goal states: If a joint state vector is published to the
/new_goal_state topic while the planner is running, then the planner
will switch its goal state to the new state. Removing any current
solutions to the old goal state and starting to search for solutions to
the new goal state...to consider: should I check the current tree to see
if any solutions already exist within it to the new goal state? Or is
that unneccessary since we will very quickly sample the goal state and
check that exact thing after changing the planner goal.