Saving state
Current state: RT-RRT* works (or appears to) for the case where we find
a plan to our initial goal state, start moving towards that initial goal
state while re-wiring the root of our tree, then get a new goal state
and change paths towards the new goal state once we reach our next joint
state while.

Issues: While we appear to rewire from the root of our tree
successfully, the path to the new goal the robot takes is not very
smooth/direct/optimal. I suspect this is because once the robot starts
moving and re-wiring the root, it stops focusing on sampling new nodes
in the environment, thus it does not have enough nodes in the tree to
create a "good" path to the goal even when all the nodes are rewired.
Need to find a good way to implement interweaving of rewiring and
sampling. Also it may help to switch to r-nearest neighbor searching
rather than k-nearest while the tree doesn't have an excessive amuont of
nodes in it. k-nearest returns up to k "neighbors" even if they aren't
near enough to connect, which seems like an innefficient use of time.
r-nearest only returns neighbors if they are close enough to connect,
which seems much more preferable (as long as there aren't an excessive
amount of nodes around you to where k-nearest would return less neigbors
than r-nearest)

Untested: behavior in the presence of obstacles for the stated
situation.

Next steps:
  - develop evaluation criteria so I can more concretely tell if my
  attempts to improve the implementation actually improve performance.
  - add RT-RRT* planner to pipeline rather than overwriting RRT* planner
  so I can use RRT* to help develop evaluation criteria.