Make initial RT-RRTstar goal a GoalState object
Previously, the initial RT-RRTstar goal was a GoalLazySamples object
that was sampling different goal states on a different thread. This is
unnecessary as the goal states I am considering are specific joint
states, so there is no need to sample different goal states. And this
creates inconsistencies as ensuing goal states are not sampled for.

Change the initial GoalLazySamples object to the simpler GoalState in
the RT-RRTstar planner.

Clean up new goal state reaction algorithm and comments in
RT-RRTstar.cpp.

Clean up goal state includes files in RT-RRTstar.cpp.

THES-35: Make OMPL goal objects consistent (Part 1)