Add RT-RRTstar ability to find shortest path quickly
Adds parameterized ability for RT-RRTstar to add the shortest path to
the goal into the tree if it is clear of obstacles. This functionality
is turned off by default since it is not really a part of the RT-RRTstar
algorithm.

The parameter "check_shortest_path" can be set to true in
`ompl_planning.yaml` to make use of this functionality.

THES-27: Implement ability to find initial path to goal state quickly
[in RT-RRTstar]