Switch RT-RRTstar r-nearest radius and clean up rewireRoot
Changes RT-RRTstar r-nearest radius to be equal to `maxDistance_`.
Otherwise the default value for `r`, calculated based on the amount of
nodes in the tree, returns few or often 0 near neighbors. This becomes
an issue in the rewireRoot method because the queue for re-wiring the
next node is populated by the near-neighbors, which runs out quickly
when the value of `r` is too small.

Clean up rewireRoot method and logging.

THES-19: Evaluate switching to r-nearest neighbors