Skip density sampling rejection if goal satisfied
Even if the area around the goal is densely sampled, we don't want to
reject a potential solution to the planning problem. Add a check on if
the goal is satisfied by the new sample before implementing density
sampling rejection.

Moves density sampling rejection up in expandTree loop so that we don't
expend extra effort of checking where best to connect the new sample to
the tree before potentially rejecting it.

THES-44: Add motion regardless of sampling density if it satisfies the
goal