THES-77: Create rerouting scheme
Create "ATTEMPT_REROUTE" state that the RTRRTstar planner will enter
when it is notified that the path between the next 2 states in the
solution path is obstructed. In this state, the planner will first
attempt to rewire the nodes in the path on the other side of the
obstruction, and if unsuccessful, attempt searching for another
solution.